fix: chapter parsing issue in ios
