+++ .flutter-plugins-dependencies
@@ -1,1 +1,1 @@
-{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_inappwebview_ios","path":"/home/fayisdev/.pub-cache/hosted/pub.dev/flutter_inappwebview_ios-1.1.2/","native_build":true,"dependencies":[],"dev_dependency":false}],"android":[{"name":"flutter_inappwebview_android","path":"/home/fayisdev/.pub-cache/hosted/pub.dev/flutter_inappwebview_android-1.1.3/","native_build":true,"dependencies":[],"dev_dependency":false}],"macos":[{"name":"flutter_inappwebview_macos","path":"/home/fayisdev/.pub-cache/hosted/pub.dev/flutter_inappwebview_macos-1.1.2/","native_build":true,"dependencies":[],"dev_dependency":false}],"linux":[],"windows":[{"name":"flutter_inappwebview_windows","path":"/home/fayisdev/.pub-cache/hosted/pub.dev/flutter_inappwebview_windows-0.6.0/","native_build":true,"dependencies":[],"dev_dependency":false}],"web":[{"name":"flutter_inappwebview_web","path":"/home/fayisdev/.pub-cache/hosted/pub.dev/flutter_inappwebview_web-1.1.2/","dependencies":[],"dev_dependency":false}]},"dependencyGraph":[{"name":"flutter_inappwebview","dependencies":["flutter_inappwebview_android","flutter_inappwebview_ios","flutter_inappwebview_macos","flutter_inappwebview_web","flutter_inappwebview_windows"]},{"name":"flutter_inappwebview_android","dependencies":[]},{"name":"flutter_inappwebview_ios","dependencies":[]},{"name":"flutter_inappwebview_macos","dependencies":[]},{"name":"flutter_inappwebview_web","dependencies":[]},{"name":"flutter_inappwebview_windows","dependencies":[]}],"date_created":"2025-03-12 19:28:34.410617","version":"3.29.0","swift_package_manager_enabled":{"ios":false,"macos":false}}
+{"info":"This is a generated file; do not edit or check into version control.","plugins":{"ios":[{"name":"flutter_inappwebview_ios","path":"/Users/fayismuhammed/.pub-cache/hosted/pub.dev/flutter_inappwebview_ios-1.1.2/","native_build":true,"dependencies":[],"dev_dependency":false}],"android":[{"name":"flutter_inappwebview_android","path":"/Users/fayismuhammed/.pub-cache/hosted/pub.dev/flutter_inappwebview_android-1.1.3/","native_build":true,"dependencies":[],"dev_dependency":false}],"macos":[{"name":"flutter_inappwebview_macos","path":"/Users/fayismuhammed/.pub-cache/hosted/pub.dev/flutter_inappwebview_macos-1.1.2/","native_build":true,"dependencies":[],"dev_dependency":false}],"linux":[],"windows":[{"name":"flutter_inappwebview_windows","path":"/Users/fayismuhammed/.pub-cache/hosted/pub.dev/flutter_inappwebview_windows-0.6.0/","native_build":true,"dependencies":[],"dev_dependency":false}],"web":[{"name":"flutter_inappwebview_web","path":"/Users/fayismuhammed/.pub-cache/hosted/pub.dev/flutter_inappwebview_web-1.1.2/","dependencies":[],"dev_dependency":false}]},"dependencyGraph":[{"name":"flutter_inappwebview","dependencies":["flutter_inappwebview_android","flutter_inappwebview_ios","flutter_inappwebview_macos","flutter_inappwebview_web","flutter_inappwebview_windows"]},{"name":"flutter_inappwebview_android","dependencies":[]},{"name":"flutter_inappwebview_ios","dependencies":[]},{"name":"flutter_inappwebview_macos","dependencies":[]},{"name":"flutter_inappwebview_web","dependencies":[]},{"name":"flutter_inappwebview_windows","dependencies":[]}],"date_created":"2025-10-26 10:03:34.625063","version":"3.35.7","swift_package_manager_enabled":{"ios":false,"macos":false}}
