+++ lib/flutter_epub_viewer.dart
@@ -2,7 +2,14 @@ library epub_viewer;
 
 export 'src/epub_viewer.dart';
 export 'src/epub_controller.dart';
-export 'src/helper.dart';
+export 'src/models/epub_chapter.dart';
+export 'src/models/epub_display_settings.dart';
+export 'src/models/epub_location.dart';
+export 'src/models/epub_search_result.dart';
+export 'src/models/epub_source.dart';
+export 'src/models/epub_text_extract_res.dart';
+export 'src/models/epub_text_selection.dart';
+export 'src/models/epub_theme.dart';
 export 'src/epub_metadata.dart';
 export 'package:flutter_inappwebview/flutter_inappwebview.dart'
     show ContextMenu, ContextMenuSettings, ContextMenuItem;
