+++ lib/src/models/epub_search_result.dart
@@ -1,0 +1,20 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'epub_search_result.g.dart';
+
+@JsonSerializable(explicitToJson: true)
+class EpubSearchResult {
+  /// The cfi string search result
+  String cfi;
+
+  /// The excerpt of the search result
+  String excerpt;
+
+  EpubSearchResult({
+    required this.cfi,
+    required this.excerpt,
+  });
+  factory EpubSearchResult.fromJson(Map<String, dynamic> json) =>
+      _$EpubSearchResultFromJson(json);
+  Map<String, dynamic> toJson() => _$EpubSearchResultToJson(this);
+}
