+++ pubspec.yaml
@@ -4,7 +4,7 @@ version: 1.2.2
 homepage: https://github.com/fayis672/epub_viewer
 
 environment:
-  sdk: '>=3.3.1 <4.0.0'
+  sdk: '>=3.8.0 <4.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
